Move all configs to wp5 base one
