+++ test/scenarios/webpack2/webpack.config.fail-sync.js
@@ -4,7 +4,7 @@
  * Fail synchronously.
  */
 const base = require("./webpack.config");
-const fail = require("../webpack4/webpack.config.fail-sync");
+const fail = require("../webpack5/webpack.config.fail-sync");
 
 module.exports = {
   ...base,
