+++ test/scenarios/webpack3/webpack.config.fail-promise.js
@@ -4,7 +4,7 @@
  * Fail promise.
  */
 const base = require("./webpack.config");
-const fail = require("../webpack4/webpack.config.fail-promise");
+const fail = require("../webpack5/webpack.config.fail-promise");
 
 module.exports = {
   ...base,
