+++ test/scenarios/webpack5/webpack.config.js
@@ -1,6 +1,5 @@
 "use strict";
 
-// TODO(WP5): Combine WP4 into this one.
 /**
  * Webpack configuration
  */
